Once again: More robust for using with STI

+++ lib/redundant_links.rb
@@ -16,7 +16,7 @@ module RedundantLinks
       if method_name.to_s =~ /\Aredundant\_linked\_\w+/ 
         association_name = method_name.to_s.gsub 'redundant_linked_', ''
         model = association_name.classify.constantize rescue nil
-        model.send(action, "scope_for_redundant_linked_#{self.class.to_s.underscore}", self)
+        model.send(action, "scope_for_redundant_linked_#{self.class.base_class.to_s.underscore}", self)
       end
     end
   end
@@ -64,47 +64,49 @@ module RedundantLinks
       options.keys.each do |klass|
         raise ArgumentError, "All keys of parameter 'options' have to be classes" unless klass.is_a?(Class)
         
-        klass.class_eval <<-EOV
-          has_many :redundant_links, :as => :to, :dependent => :delete_all
+        unless klass == self
+          klass.class_eval <<-EOV
+            has_many :redundant_links, :as => :to, :dependent => :delete_all
 
-          # Update links, if master record was changed
-          after_update :update_redundant_links_for_#{self.to_s.tableize}
+            # Update links, if master record was changed
+            after_update :update_redundant_links_for_#{self.to_s.tableize}
 
-          def update_redundant_links_for_#{self.to_s.tableize}
-            first_record = redundant_linked_#{self.to_s.tableize}.first
+            def update_redundant_links_for_#{self.to_s.tableize}
+              first_record = redundant_linked_#{self.to_s.tableize}.first
             
-            changed = false
-            if targets = #{self}.redundant_links_options[#{klass}]
-              targets.each do |target|              
-                if target_object = self.send(target)
-                  if first_record
-                    changed = true if not RedundantLink.exists?(:from_id => first_record.id, :from_type => first_record.class.base_class.to_s, :to_id => target_object.id, :to_type => target_object.class.base_class.to_s)
+              changed = false
+              if targets = #{self}.redundant_links_options[#{klass}]
+                targets.each do |target|              
+                  if target_object = self.send(target)
+                    if first_record
+                      changed = true if not RedundantLink.exists?(:from_id => first_record.id, :from_type => first_record.class.base_class.to_s, :to_id => target_object.id, :to_type => target_object.class.base_class.to_s)
+                    else
+                      changed = true
+                    end
                   else
-                    changed = true
+                    if first_record
+                      changed = true if RedundantLink.exists?(:from_id => first_record.id, :from_type => first_record.class.base_class.to_s)
+                    end
                   end
-                else
-                  if first_record
-                    changed = true if RedundantLink.exists?(:from_id => first_record.id, :from_type => first_record.class.base_class.to_s)
-                  end
-                end
               
-                break if changed
+                  break if changed
+                end
               end
-            end
             
-            if changed
-              self.redundant_linked_#{self.to_s.tableize}.each do |record|
-                record.send :update_redundant_links
+              if changed
+                self.redundant_linked_#{self.to_s.tableize}.each do |record|
+                  record.send :update_redundant_links
+                end
               end
             end
-          end
-        EOV
+          EOV
+        end
         
         # In the detail class: Build a class method to get a scope, e.g. "scope_for_redundant_linked_customer(record)"
         self.class_eval <<-EOV
-          def self.scope_for_redundant_linked_#{klass.to_s.underscore}(record)
+          def self.scope_for_redundant_linked_#{klass.base_class.to_s.underscore}(record)
             scoped :joins => "INNER JOIN redundant_links ON (redundant_links.to_id     = \#{record.id.is_a?(String) ? '"' + record.id + '"' : record.id }
-                                                         AND redundant_links.to_type   = '#{klass.base_class}'
+                                                         AND redundant_links.to_type   = '\#{record.class.base_class}'
                                                          AND redundant_links.from_id   = #{self.to_s.tableize}.id
                                                          AND redundant_links.from_type = '#{self}')"
           end
